# perf(resolve): only read license/ deprecation when used in rule set