--- a/src/extract/resolve/determine-dependency-types.js
+++ b/src/extract/resolve/determine-dependency-types.js
@@ -104,6 +104,7 @@ function determineNodeModuleDependencyTypes(
   );
 
   if (
+    pResolveOptions.resolveDeprecations &&
     localNpmHelpers.dependencyIsDeprecated(
       pModuleName,
       pFileDirectory,