--- a/src/extract/resolve/resolve-helpers.js
+++ b/src/extract/resolve/resolve-helpers.js
@@ -3,14 +3,16 @@ const localNpmHelpers = require("./local-npm-helpers");
 module.exports = {
   addLicenseAttribute(pModuleName, pBaseDirectory, pResolveOptions) {
     let lReturnValue = {};
-    const lLicense = localNpmHelpers.getLicense(
-      pModuleName,
-      pBaseDirectory,
-      pResolveOptions
-    );
+    if (pResolveOptions.resolveLicenses) {
+      const lLicense = localNpmHelpers.getLicense(
+        pModuleName,
+        pBaseDirectory,
+        pResolveOptions
+      );
 
-    if (Boolean(lLicense)) {
-      lReturnValue.license = lLicense;
+      if (Boolean(lLicense)) {
+        lReturnValue.license = lLicense;
+      }
     }
     return lReturnValue;
   },