… src/graph-utl/find-rule-by-name.js
… test/graph-utl/find-rule-by-name.spec.js