## Description, Motivation and Context

When 'normalising' the options, it determines whether a license check and/ or a deprecation check against external dependencies is necessary. If it isn't necessary skipping these checks saves a significant amount of time (averages over consecutive runs on a 2,6 GHz Quad-Core Intel Core i7 with an SSD with APFS - macOS 10.15.7); 
- For dependency-cruiser's own scan it saved ~1.5s (for the license check) and ~0.3s (for the deprecation check) on a total run time taking ~3.2s (for 395 modules, 939 dependencies) before these improvements.
- For yarn 'berry' it saved ~2s (license check + deprecation check) on a total run time originaly taking ~9.5s (for 745 modules, 1262 dependencies using ts -> ast ) before these improvements.

## How Has This Been Tested?

- [x] automated non-regression tests
- [x] green ci
- [x] additional unit tests
 
## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [x] Refactor (non-breaking change which fixes an issue without changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).